requirements.txt
yt-dlp>=2024.1.0
google-generativeai>=0.8.0
requests>=2.31.0
